Split the main get-triggers file into separates + better promises
