+++ summerfi-api/get-triggers-function/src/helpers/logger.ts
@@ -1,0 +1,3 @@
+import { Logger } from '@aws-lambda-powertools/logger'
+
+export const logger = new Logger({ serviceName: 'get-triggers-function' })
